Revert "Added Friend Request Feature"
