+++ public/css/main.css
@@ -25,7 +25,7 @@ input{
     padding: 10px;
 }
 
-#btn, .btn{
+#btn{
     font-size: 30px;
     background-color: #DDA0DD;
     color: #E0FFFF;
