+++ views/layouts/main.handlebars
@@ -6,12 +6,7 @@
     <link rel="stylesheet" href="/public/css/main.css">
 </head>
 <body>
-    <header id="nav">
-        <a href="/feed">Feed</a>
-        <a href="/logout">Logout</a>
-        <a href="/editprofile">Edit profile</a>
-        <a href="/friendrequests">Friend Requests</a>
-    </header>
+    
     {{{body}}}
 
     <script src="https://code.jquery.com/jquery-2.2.4.min.js" integrity="sha256-BbhdlvQf/xTY9gja0Dq3HiwQF8LaCRTXxZKRutelT44="
