+++ views/render/editProfile.handlebars
@@ -44,4 +44,8 @@
     <div class="error">
         <label for="errorLabel" id="errorLabel">{{msg}}</label>
     </div>
+    <footer>
+        <a href="/feed">Feed</a>
+        <a href="/logout">Logout</a>
+    </footer>
 </div>
