--- views/render/friendRequests.handlebars
@@ -1,8 +1,0 @@
-<div id="friendRequests">
-
-    <h1 id="userID">{{currUser}}</h1>
-    
-    <script src="https://code.jquery.com/jquery-2.2.4.min.js" integrity="sha256-BbhdlvQf/xTY9gja0Dq3HiwQF8LaCRTXxZKRutelT44="
-    crossorigin="anonymous"></script>
-    <script src="/public/js/friendRequests.js"></script>
-</div>
