+++ views/render/home.handlebars
@@ -1,4 +1,4 @@
-<div id="home">
+<div>
     <h1>
         Welcome
     </h1>
