+++ views/render/userInfo.handlebars
@@ -4,12 +4,11 @@
         {{data.firstName}} {{data.lastName}}
     </h1>
 
-    <button class="btn" id="friendReqBtn">Add Friend</button>
-
     <div id="userInfoCat">
         <a href="" id="userBioClick">User Information</a>
         <a href="" id="userPostsClick">User Posts</a>
         <a href="" id="userLikedClick">Liked by User</a>
+        <a href="/editprofile" id="editProfileClick">Edit profile</a>
     </div>
 
     <div id="userBio">
@@ -23,7 +22,9 @@
     <div id="userPosts">
 
     </div>
-    <script src="https://code.jquery.com/jquery-2.2.4.min.js" integrity="sha256-BbhdlvQf/xTY9gja0Dq3HiwQF8LaCRTXxZKRutelT44="
-    crossorigin="anonymous"></script>
-    <script src="/public/js/userInfo.js"></script>
+
+    <footer>
+        <a href="/feed">Feed</a>
+        <a href="/logout">Logout</a>
+    </footer>
 </div>
